% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEPEst0624-package.R
\docType{package}
\name{DEEPEst0624-package}
\alias{DEEPEst0624-package}
\alias{DEEPEst0624}
\title{The 'DEEPEst0624' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
